Add abstractions for classes that have a logger
